base gemm template succ

+++ benchmarks/MatrixMulCUDA13/gemm_base.h
@@ -10,7 +10,7 @@ struct iBaseGemm {
 
   // virtual void GetMetric(Metric& metric, int m, int n, int k) = 0;
 
-  virtual void Launch(float* C, const float* A, const float* B, int M, int N,
+  virtual void Launch(__half* C, const __half* A, const __half* B, int M, int N,
                       int K) = 0;
 
   virtual void Dump(std::ostream& os) = 0;
